Fix a compiler warning about checking differing enum types One of the enum types is actually a subset of the other, and share the same integer values for compatible enumation values

+++ OpenAL32/alBuffer.c
@@ -1514,7 +1514,7 @@ static ALenum LoadData(ALbuffer *ALBuf, ALuint freq, ALenum NewFormat, ALsizei s
     NewChannels = ChannelsFromFmt(DstChannels);
     NewBytes = BytesFromFmt(DstType);
 
-    assert(SrcChannels == DstChannels);
+    assert((long)SrcChannels == (long)DstChannels);
 
     if(SrcType == UserFmtIMA4)
     {
